Organized unit tests, added unit tests for GetAllFlags and GetFlagById. [#136582371]
