--- RevaShare/Ray-Unit-Tests.playlist
@@ -1,1 +1,0 @@
-<Playlist Version="1.0"><Add Test="RevaShare.Test.RideCRUD.UpdateRide_Test" /><Add Test="RevaShare.Test.UserCRUD.GetUserByUsername_Test" /><Add Test="RevaShare.Test.UserCRUD.AddAdminDeleteUser_Test" /><Add Test="RevaShare.Test.UserCRUD.RegisterUserDeleteUser_Test" /><Add Test="RevaShare.Test.RideCRUD.ListRidesAtApartment_Test" /><Add Test="RevaShare.Test.RideCRUD.AddRideRemoveRide_Test" /><Add Test="RevaShare.Test.UserCRUD.UpdateUser_Test" /><Add Test="RevaShare.Test.RideCRUD.GetAllRides_Test" /><Add Test="RevaShare.Test.UserCRUD.GetAllUsers_Test" /></Playlist>
