+++ RevaShare/RevaShare.DataClient/RevaShare.DataClient.csproj
@@ -151,6 +151,7 @@
   <ItemGroup>
     <Content Include="packages.config" />
     <None Include="Properties\PublishProfiles\RevaShare.pubxml" />
+    <None Include="Properties\PublishProfiles\RevaShare_Data-Access_Jason-EC2.pubxml" />
     <None Include="Properties\PublishProfiles\TeamDB-RevaShare.pubxml" />
     <None Include="Web.Debug.config">
       <DependentUpon>Web.config</DependentUpon>
