+++ RevaShare/RevaShare.Test/App.config
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <configSections>
     <!-- For more information on Entity Framework configuration, visit http://go.microsoft.com/fwlink/?LinkID=237468 -->
@@ -11,10 +11,8 @@
     </providers>
   </configProtectedData>
   <connectionStrings>
-     <add name="RevaShareDBEntities" connectionString="metadata=res://*/RSModel.csdl|res://*/RSModel.ssdl|res://*/RSModel.msl;provider=System.Data.SqlClient;provider connection string=&quot;data source=sqltraining.c8xdjhpxfhp5.us-east-1.rds.amazonaws.com;initial catalog=RevashareDB_DEV;persist security info=True;user id=sqlshare;password=123456;MultipleActiveResultSets=True;App=EntityFramework&quot;"
-    providerName="System.Data.EntityClient" />
-    <add name="IdentityDBEntities" connectionString="Data Source=sqltraining.cdtgqwlp4y7y.us-east-1.rds.amazonaws.com;Initial Catalog=RevaShareIdentityDB;User id=sqlroot;Password=Pass1234; MultipleActiveResultSets=true;"
-      providerName="System.Data.SqlClient" />
+     <add name="RevaShareDBEntities" connectionString="metadata=res://*/RSModel.csdl|res://*/RSModel.ssdl|res://*/RSModel.msl;provider=System.Data.SqlClient;provider connection string=&quot;data source=sqltraining.c8xdjhpxfhp5.us-east-1.rds.amazonaws.com;initial catalog=RevashareDB_DEV;persist security info=True;user id=sqlshare;password=123456;MultipleActiveResultSets=True;App=EntityFramework&quot;" providerName="System.Data.EntityClient" />
+    <add name="IdentityDBEntities" connectionString="Data Source=sqltraining.cdtgqwlp4y7y.us-east-1.rds.amazonaws.com;Initial Catalog=RevaShareIdentityDB;User id=sqlroot;Password=Pass1234; MultipleActiveResultSets=true;" providerName="System.Data.SqlClient" />
   </connectionStrings>
   <entityFramework>
     <!--<defaultConnectionFactory type="System.Data.Entity.Infrastructure.LocalDbConnectionFactory, EntityFramework">
