+++ RevaShare/RevaShare.Test/packages.config
@@ -9,6 +9,5 @@
   <package id="xunit.core" version="2.1.0" targetFramework="net452" />
   <package id="xunit.extensibility.core" version="2.1.0" targetFramework="net452" />
   <package id="xunit.extensibility.execution" version="2.1.0" targetFramework="net452" />
-  <package id="xunit.runner.console" version="2.1.0" targetFramework="net452" />
   <package id="xunit.runner.visualstudio" version="2.1.0" targetFramework="net452" />
 </packages>
